+++ app/src/main/java/com/example/swapit/data/datasource/remote/dto/response/shopping/detail/ShoppingDetailImageResponse.kt
@@ -4,6 +4,6 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class ShoppingDetailImageResponse(
-    val imagesId: Int,
+    val imagesId: Long,
     val imageUrl: String,
 )
