+++ app/src/main/java/com/example/swapit/domain/model/shopping/ShoppingProduct.kt
@@ -1,12 +1,12 @@
 package com.example.swapit.domain.model.shopping
 
 data class ShoppingProduct(
-    val goodsId: Int,
+    val goodsId: Long,
     val title: String,
-    val price: Int,
+    val price: Long,
     val category: String,
     val imageUrl: String?,
     val placeName: String,
-    val viewCount: Int,
+    val viewCount: Long,
     val createdAt: String,
 )
