+++ app/src/main/java/com/example/swapit/domain/model/shopping/ShoppingProductResults.kt
@@ -5,6 +5,6 @@ import com.example.swapit.data.datasource.remote.dto.response.shopping.ShoppingP
 data class ShoppingProductResults(
     val goodsList: List<ShoppingProductResponse>,
     val hasNext: Boolean,
-    val lastCursorId: Int,
-    val count: Int,
+    val lastCursorId: Long,
+    val count: Long,
 )
