+++ app/src/main/java/com/example/swapit/domain/model/shopping/detail/ShoppingDetailImage.kt
@@ -1,6 +1,6 @@
 package com.example.swapit.domain.model.shopping.detail
 
 data class ShoppingDetailImage(
-    val imagesId: Int,
+    val imagesId: Long,
     val imageUrl: String,
 )
